+++ R/datasummary_correlation.R
@@ -18,6 +18,7 @@
 #'     `datasummary_correlation_format` can often be useful for formatting the
 #'     output of custom correlation functions.
 #' }
+#' @template citation
 #' @template options
 #' @param ... other parameters are passed through to the table-making
 #'     packages.