+++ man-roxygen/citation.R
@@ -1,0 +1,4 @@
+#'
+#' @references
+#' Arel-Bundock V (2022). “modelsummary: Data and Model Summaries in R.” _Journal of Statistical Software_, *103*(1), 1-23. doi:10.18637/jss.v103.i01 <https://doi.org/10.18637/jss.v103.i01>.'
+#'